azure-functions/chat-proxy/requirements.txt
azure-functions-core-tools
azure-functions
httpx
